Logic act pomodoro task timer and total tasks completed session timer completed. huge change: not using link table, but store it directly to their corresponding tables.
fixing bug in push notification, the push notification for hotkey has to only occur when the setting is enabled.

+++ ExtendedPomodoro/ViewModels/TaskDomainViewModel.cs
@@ -15,6 +15,6 @@ namespace ExtendedPomodoro.ViewModels
         int ActPomodoro,
         string CreatedAt,
         int TaskStatus,
-        int TimeSpentInMinutes
+        double TimeSpentInMinutes
         );
 }
